divide bottom nav
